refactored duplicates vars
