+++ pkg/apis/events/v1alpha1/test_helper.go
@@ -1,7 +1,7 @@
-package v1beta1
+package v1alpha1
 
 import (
-	duckv1beta1 "github.com/google/knative-gcp/pkg/apis/duck/v1beta1"
+	duckv1alpha1 "github.com/google/knative-gcp/pkg/apis/duck/v1alpha1"
 	"github.com/google/knative-gcp/pkg/apis/events"
 	duckv1 "knative.dev/pkg/apis/duck/v1"
 )
@@ -9,18 +9,18 @@ import (
 // These variables are used to create a 'complete' version of Source where every field is
 // filled in.
 var (
-	CompleteIdentitySpec = duckv1beta1.IdentitySpec{
+	CompleteIdentitySpec = duckv1alpha1.IdentitySpec{
 		ServiceAccountName: "k8sServiceAccount",
 	}
 
-	completePubSubSpec = duckv1beta1.PubSubSpec{
+	completePubSubSpec = duckv1alpha1.PubSubSpec{
 		SourceSpec:   events.CompleteSourceSpec,
 		IdentitySpec: CompleteIdentitySpec,
 		Secret:       events.CompleteSecret,
 		Project:      "project",
 	}
 
-	completeIdentityStatus = duckv1beta1.IdentityStatus{
+	completeIdentityStatus = duckv1alpha1.IdentityStatus{
 		Status: duckv1.Status{
 			ObservedGeneration: 7,
 			Conditions: duckv1.Conditions{
@@ -33,7 +33,7 @@ var (
 		ServiceAccountName: "serviceAccountName",
 	}
 
-	completePubSubStatus = duckv1beta1.PubSubStatus{
+	completePubSubStatus = duckv1alpha1.PubSubStatus{
 		IdentityStatus: completeIdentityStatus,
 		SinkURI:        &events.CompleteURL,
 		CloudEventAttributes: []duckv1.CloudEventAttributes{
